+++ src/features/pmicControl/npm/pmic1300Device.test.ts
@@ -621,7 +621,7 @@ describe('PMIC 1300', () => {
             });
 
             test('Request startAdcSample', () => {
-                pmic.startAdcSample(2000);
+                pmic.startAdcSample(2000, 1000);
 
                 expect(mockEnqueueRequest).toBeCalledTimes(1);
                 expect(mockEnqueueRequest).toBeCalledWith(
@@ -637,7 +637,7 @@ describe('PMIC 1300', () => {
 
                 expect(mockEnqueueRequest).toBeCalledTimes(1);
                 expect(mockEnqueueRequest).toBeCalledWith(
-                    'npm_adc sample 1000 0',
+                    'npm_adc sample 0',
                     expect.anything(),
                     undefined,
                     true
@@ -743,7 +743,7 @@ describe('PMIC 1300', () => {
                     });
 
                     // turn off charging
-                    expect(mockEnqueueRequest).toBeCalledTimes(2);
+                    expect(mockEnqueueRequest).toBeCalledTimes(3);
                     expect(mockEnqueueRequest).nthCalledWith(
                         1,
                         `npmx charger module charger set 0`,
@@ -753,6 +753,13 @@ describe('PMIC 1300', () => {
                     );
                     expect(mockEnqueueRequest).nthCalledWith(
                         2,
+                        `npm_adc sample 1000 2000`,
+                        expect.anything(),
+                        undefined,
+                        true
+                    );
+                    expect(mockEnqueueRequest).nthCalledWith(
+                        3,
                         `npmx charger termination_voltage normal set 3200`,
                         expect.anything(),
                         undefined,
@@ -773,7 +780,7 @@ describe('PMIC 1300', () => {
                     });
 
                     // turn off charging
-                    expect(mockEnqueueRequest).toBeCalledTimes(2);
+                    expect(mockEnqueueRequest).toBeCalledTimes(3);
                     expect(mockEnqueueRequest).nthCalledWith(
                         1,
                         `npmx charger module charger set 0`,
@@ -783,6 +790,13 @@ describe('PMIC 1300', () => {
                     );
                     expect(mockEnqueueRequest).nthCalledWith(
                         2,
+                        `npm_adc sample 1000 2000`,
+                        expect.anything(),
+                        undefined,
+                        true
+                    );
+                    expect(mockEnqueueRequest).nthCalledWith(
+                        3,
                         `npmx charger charger_current set 32`,
                         expect.anything(),
                         undefined,
@@ -864,9 +878,17 @@ describe('PMIC 1300', () => {
             )('Set setChargerEnabled %p', async ({ index, enabled }) => {
                 await pmic.setChargerEnabled(index, enabled);
 
-                expect(mockEnqueueRequest).toBeCalledTimes(1);
-                expect(mockEnqueueRequest).toBeCalledWith(
+                expect(mockEnqueueRequest).toBeCalledTimes(2);
+                expect(mockEnqueueRequest).nthCalledWith(
+                    1,
                     `npmx charger module charger set ${enabled ? '1' : '0'}`,
+                    expect.anything(),
+                    undefined,
+                    true
+                );
+                expect(mockEnqueueRequest).nthCalledWith(
+                    2,
+                    `npm_adc sample ${enabled ? '500' : '1000'} 2000`,
                     expect.anything(),
                     undefined,
                     true
@@ -1466,7 +1488,7 @@ describe('PMIC 1300', () => {
                     });
 
                     // turn off charging
-                    expect(mockEnqueueRequest).toBeCalledTimes(3);
+                    expect(mockEnqueueRequest).toBeCalledTimes(4);
                     expect(mockEnqueueRequest).nthCalledWith(
                         1,
                         `npmx charger module charger set 0`,
@@ -1476,6 +1498,13 @@ describe('PMIC 1300', () => {
                     );
                     expect(mockEnqueueRequest).nthCalledWith(
                         2,
+                        `npm_adc sample 1000 2000`,
+                        expect.anything(),
+                        undefined,
+                        true
+                    );
+                    expect(mockEnqueueRequest).nthCalledWith(
+                        3,
                         `npmx charger termination_voltage normal set 3200`,
                         expect.anything(),
                         undefined,
@@ -1484,7 +1513,7 @@ describe('PMIC 1300', () => {
 
                     // Refresh data due to error
                     expect(mockEnqueueRequest).nthCalledWith(
-                        3,
+                        4,
                         `npmx charger termination_voltage normal get`,
                         expect.anything(),
                         undefined,
@@ -1551,7 +1580,7 @@ describe('PMIC 1300', () => {
                     });
 
                     // turn off charging
-                    expect(mockEnqueueRequest).toBeCalledTimes(3);
+                    expect(mockEnqueueRequest).toBeCalledTimes(4);
                     expect(mockEnqueueRequest).nthCalledWith(
                         1,
                         `npmx charger module charger set 0`,
@@ -1561,6 +1590,13 @@ describe('PMIC 1300', () => {
                     );
                     expect(mockEnqueueRequest).nthCalledWith(
                         2,
+                        `npm_adc sample 1000 2000`,
+                        expect.anything(),
+                        undefined,
+                        true
+                    );
+                    expect(mockEnqueueRequest).nthCalledWith(
+                        3,
                         `npmx charger charger_current set 32`,
                         expect.anything(),
                         undefined,
@@ -1569,7 +1605,7 @@ describe('PMIC 1300', () => {
 
                     // Refresh data due to error
                     expect(mockEnqueueRequest).nthCalledWith(
-                        3,
+                        4,
                         `npmx charger charger_current get`,
                         expect.anything(),
                         undefined,
@@ -1635,7 +1671,7 @@ describe('PMIC 1300', () => {
                     });
 
                     // turn off charging
-                    expect(mockEnqueueRequest).toBeCalledTimes(3);
+                    expect(mockEnqueueRequest).toBeCalledTimes(4);
                     expect(mockEnqueueRequest).nthCalledWith(
                         1,
                         `npmx charger module charger set 0`,
@@ -1645,6 +1681,13 @@ describe('PMIC 1300', () => {
                     );
                     expect(mockEnqueueRequest).nthCalledWith(
                         2,
+                        `npm_adc sample 1000 2000`,
+                        expect.anything(),
+                        undefined,
+                        true
+                    );
+                    expect(mockEnqueueRequest).nthCalledWith(
+                        3,
                         `npmx charger trickle set 2500`,
                         expect.anything(),
                         undefined,
@@ -1653,7 +1696,7 @@ describe('PMIC 1300', () => {
 
                     // Refresh data due to error
                     expect(mockEnqueueRequest).nthCalledWith(
-                        3,
+                        4,
                         `npmx charger trickle get`,
                         expect.anything(),
                         undefined,
@@ -1719,7 +1762,7 @@ describe('PMIC 1300', () => {
                     });
 
                     // turn off charging
-                    expect(mockEnqueueRequest).toBeCalledTimes(3);
+                    expect(mockEnqueueRequest).toBeCalledTimes(4);
                     expect(mockEnqueueRequest).nthCalledWith(
                         1,
                         `npmx charger module charger set 0`,
@@ -1729,6 +1772,13 @@ describe('PMIC 1300', () => {
                     );
                     expect(mockEnqueueRequest).nthCalledWith(
                         2,
+                        `npm_adc sample 1000 2000`,
+                        expect.anything(),
+                        undefined,
+                        true
+                    );
+                    expect(mockEnqueueRequest).nthCalledWith(
+                        3,
                         `npmx charger termination_current set 10`,
                         expect.anything(),
                         undefined,
@@ -1737,7 +1787,7 @@ describe('PMIC 1300', () => {
 
                     // Refresh data due to error
                     expect(mockEnqueueRequest).nthCalledWith(
-                        3,
+                        4,
                         `npmx charger termination_current get`,
                         expect.anything(),
                         undefined,